Scala Meta macro annotation for tagged types - refactored
